+++ dev/src/webs_lib/mod.rs
@@ -6,8 +6,8 @@ use rand::rngs::ThreadRng;
 use rand::Rng;
 
 pub const WINDOW_EDGE: f64 = 800.0;
-pub const WINDOW_EDGE_HALF: f64 = WINDOW_EDGE / 2.0;
-pub const WINDOW_EDGE_HALF_MINUS: f64 = -WINDOW_EDGE_HALF;
+pub const WINDOW_EDGE_HALF: f32 = (WINDOW_EDGE as f32) / 2.0;
+pub const WINDOW_EDGE_HALF_MINUS: f32 = -WINDOW_EDGE_HALF;
 
 pub const ANTI_ALIAS: u8 = 4;
 
@@ -36,8 +36,8 @@ const INTERSECTIONS_CAP: usize = 16;
 pub const NODES_LIMIT: usize = NODES_CAP - 2;
 pub const EDGES_LIMIT: usize = EDGES_CAP - 3;
 
-pub const POINT_RNG_UPPER: f32 = WINDOW_EDGE_HALF as f32;
-pub const POINT_RNG_LOWER: f32 = WINDOW_EDGE_HALF_MINUS as f32;
+pub const POINT_RNG_UPPER: f32 = WINDOW_EDGE_HALF;
+pub const POINT_RNG_LOWER: f32 = WINDOW_EDGE_HALF_MINUS;
 
 pub const POINT_DRAG: f32 = 0.0025;
 pub const NEIGHBOR_DISTANCE_SQUARED: f32 = 100.0;
